Added JWT token based authentication

+++ src/main/java/com/sivalabs/devzone/links/web/controller/LinkController.java
@@ -1,15 +1,18 @@
 package com.sivalabs.devzone.links.web.controller;
 
 import com.sivalabs.devzone.common.exceptions.ResourceNotFoundException;
-import com.sivalabs.devzone.links.entities.Link;
 import com.sivalabs.devzone.links.models.LinkDTO;
 import com.sivalabs.devzone.links.models.LinksDTO;
 import com.sivalabs.devzone.links.services.LinkService;
 import com.sivalabs.devzone.users.entities.User;
+import com.sivalabs.devzone.users.services.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.eclipse.microprofile.jwt.JsonWebToken;
 
+import javax.annotation.security.PermitAll;
+import javax.annotation.security.RolesAllowed;
 import javax.validation.Valid;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -23,8 +26,10 @@ import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.UriInfo;
 import java.net.URI;
+import java.util.Objects;
 
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
@@ -33,8 +38,11 @@ import java.net.URI;
 @Slf4j
 public class LinkController {
     private final LinkService linkService;
+    private final UserService userService;
+    private final JsonWebToken jwt;
 
     @GET
+    @PermitAll
     public LinksDTO findAllLinks(
             @QueryParam("page") @DefaultValue("1") int page,
             @QueryParam("query") @DefaultValue("") String query,
@@ -54,6 +62,7 @@ public class LinkController {
 
     @GET
     @Path("/{id}")
+    @PermitAll
     public LinkDTO getLink(@PathParam("id") Long id) {
         return linkService
                 .getLinkById(id)
@@ -61,8 +70,12 @@ public class LinkController {
     }
 
     @POST
-    public Response createLink(@Valid LinkDTO link, @Context UriInfo uriInfo) {
-        link.setCreatedUserId(1L);
+    @RolesAllowed({ "ROLE_USER", "ROLE_ADMIN" })
+    public Response createLink(@Valid LinkDTO link, @Context UriInfo uriInfo,
+                               @Context SecurityContext ctx) {
+        String email = jwt.getClaim("email");
+        User user = userService.getUserByEmail(email).orElseThrow();
+        link.setCreatedUserId(user.getId());
         LinkDTO linkDTO = linkService.createLink(link);
         URI uri = uriInfo.getAbsolutePathBuilder().path(Long.toString(linkDTO.getId())).build();
         return Response.created(uri).entity(linkDTO).build();
@@ -70,10 +83,16 @@ public class LinkController {
 
     @DELETE
     @Path("/{id}")
-    public Response deleteLink(@PathParam("id") Long id) {
+    @RolesAllowed({ "ROLE_USER", "ROLE_ADMIN" })
+    public Response deleteLink(@PathParam("id") Long id,
+                               @Context SecurityContext ctx) {
         LinkDTO link = linkService.getLinkById(id).orElseThrow();
-        //this.checkPrivilege(id, link, loginUser);
-        linkService.deleteLink(link.getId());
-        return Response.noContent().build();
+        String email = jwt.getClaim("email");
+        User user = userService.getUserByEmail(email).orElseThrow();
+        if(ctx.isUserInRole("ROLE_ADMIN") || Objects.equals(link.getCreatedUserId(), user.getId())) {
+            linkService.deleteLink(link.getId());
+            return Response.noContent().build();
+        }
+        return Response.status(Response.Status.FORBIDDEN).build();
     }
 }
